Fix view other social user's profile page
